add: rename project to rockervsc
