+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## python_template
+## rockervsc
 
 ## [0.0.0]
